Fix UndefVar error and document LLA->ECEF conversion convention
Before this fix, converting from ECEF to LLA coordinates with
points very near to the centre of the ellipsoid and with a small
value of z (e.g., (x, y, z) = (4e4, 0, 0) for WGS84) would throw
an UndefVarError.  With this fix, the same LLA point is returned
as is done by GeographicLib.  However, in this case multiple
solutions are present and this fix also documents the choice made.

- Change variable name to avoid UndefVarError
- Add test based off GeographicLib values
- Add explanation of which of several solutions are chosen in the
  ECEF->LLA conversion to the README.md file.